+++ chunk_reader.go
@@ -45,7 +45,7 @@ func getBooleanValuesDecoder(pageEncoding parquet.Encoding, dictValues []interfa
 	case parquet.Encoding_RLE:
 		return &booleanRLEDecoder{}, nil
 	case parquet.Encoding_RLE_DICTIONARY:
-		return &dictDecoder{values: dictValues}, nil
+		return &dictDecoder{uniqueValues: dictValues}, nil
 	default:
 		return nil, errors.Errorf("unsupported encoding %s for boolean", pageEncoding)
 	}
@@ -60,7 +60,7 @@ func getByteArrayValuesDecoder(pageEncoding parquet.Encoding, dictValues []inter
 	case parquet.Encoding_DELTA_BYTE_ARRAY:
 		return &byteArrayDeltaDecoder{}, nil
 	case parquet.Encoding_RLE_DICTIONARY:
-		return &dictDecoder{values: dictValues}, nil
+		return &dictDecoder{uniqueValues: dictValues}, nil
 	default:
 		return nil, errors.Errorf("unsupported encoding %s for binary", pageEncoding)
 	}
@@ -73,7 +73,7 @@ func getFixedLenByteArrayValuesDecoder(pageEncoding parquet.Encoding, len int, d
 	case parquet.Encoding_DELTA_BYTE_ARRAY:
 		return &byteArrayDeltaDecoder{}, nil
 	case parquet.Encoding_RLE_DICTIONARY:
-		return &dictDecoder{values: dictValues}, nil
+		return &dictDecoder{uniqueValues: dictValues}, nil
 	default:
 		return nil, errors.Errorf("unsupported encoding %s for fixed_len_byte_array(%d)", pageEncoding, len)
 	}
@@ -86,7 +86,7 @@ func getInt32ValuesDecoder(pageEncoding parquet.Encoding, typ *parquet.SchemaEle
 	case parquet.Encoding_DELTA_BINARY_PACKED:
 		return &int32DeltaBPDecoder{}, nil
 	case parquet.Encoding_RLE_DICTIONARY:
-		return &dictDecoder{values: dictValues}, nil
+		return &dictDecoder{uniqueValues: dictValues}, nil
 	default:
 		return nil, errors.Errorf("unsupported encoding %s for int32", pageEncoding)
 	}
@@ -99,7 +99,7 @@ func getInt64ValuesDecoder(pageEncoding parquet.Encoding, typ *parquet.SchemaEle
 	case parquet.Encoding_DELTA_BINARY_PACKED:
 		return &int64DeltaBPDecoder{}, nil
 	case parquet.Encoding_RLE_DICTIONARY:
-		return &dictDecoder{values: dictValues}, nil
+		return &dictDecoder{uniqueValues: dictValues}, nil
 	default:
 		return nil, errors.Errorf("unsupported encoding %s for int64", pageEncoding)
 	}
@@ -128,7 +128,7 @@ func getValuesDecoder(pageEncoding parquet.Encoding, typ *parquet.SchemaElement,
 		case parquet.Encoding_PLAIN:
 			return &floatPlainDecoder{}, nil
 		case parquet.Encoding_RLE_DICTIONARY:
-			return &dictDecoder{values: dictValues}, nil
+			return &dictDecoder{uniqueValues: dictValues}, nil
 		}
 
 	case parquet.Type_DOUBLE:
@@ -136,7 +136,7 @@ func getValuesDecoder(pageEncoding parquet.Encoding, typ *parquet.SchemaElement,
 		case parquet.Encoding_PLAIN:
 			return &doublePlainDecoder{}, nil
 		case parquet.Encoding_RLE_DICTIONARY:
-			return &dictDecoder{values: dictValues}, nil
+			return &dictDecoder{uniqueValues: dictValues}, nil
 		}
 
 	case parquet.Type_INT32:
@@ -150,7 +150,7 @@ func getValuesDecoder(pageEncoding parquet.Encoding, typ *parquet.SchemaElement,
 		case parquet.Encoding_PLAIN:
 			return &int96PlainDecoder{}, nil
 		case parquet.Encoding_RLE_DICTIONARY:
-			return &dictDecoder{values: dictValues}, nil
+			return &dictDecoder{uniqueValues: dictValues}, nil
 		}
 
 	default:
@@ -168,10 +168,9 @@ func createDataReader(r io.Reader, codec parquet.CompressionCodec, compressedSiz
 	return newBlockReader(r, codec, compressedSize, uncompressedSize)
 }
 
-func readPages(ctx context.Context, r *offsetReader, col *Column, chunkMeta *parquet.ColumnMetaData, dDecoder, rDecoder getLevelDecoder) ([]pageReader, error) {
+func readPages(ctx context.Context, r *offsetReader, col *Column, chunkMeta *parquet.ColumnMetaData, dDecoder, rDecoder getLevelDecoder) (pages []pageReader, useDict bool, err error) {
 	var (
 		dictPage *dictPageReader
-		pages    []pageReader
 	)
 
 	for {
@@ -180,35 +179,34 @@ func readPages(ctx context.Context, r *offsetReader, col *Column, chunkMeta *par
 		}
 		ph := &parquet.PageHeader{}
 		if err := readThrift(ctx, ph, r); err != nil {
-			return nil, err
+			return nil, false, err
 		}
 
 		if ph.Type == parquet.PageType_DICTIONARY_PAGE {
 			if dictPage != nil {
-				return nil, errors.New("there should be only one dictionary")
+				return nil, false, errors.New("there should be only one dictionary")
 			}
 			p := &dictPageReader{}
 			de, err := getDictValuesDecoder(col.Element())
 			if err != nil {
-				return nil, err
+				return nil, false, err
 			}
 			if err := p.init(de); err != nil {
-				return nil, err
+				return nil, false, err
 			}
 
-			// re-use the value dictionary store
-			p.values = col.getColumnStore().values.values
 			if err := p.read(r, ph, chunkMeta.Codec); err != nil {
-				return nil, err
+				return nil, false, err
 			}
 
 			dictPage = p
+
 			// Go to the next data Page
 			// if we have a DictionaryPageOffset we should return to DataPageOffset
 			if chunkMeta.DictionaryPageOffset != nil {
 				if *chunkMeta.DictionaryPageOffset != r.offset {
 					if _, err := r.Seek(chunkMeta.DataPageOffset, io.SeekStart); err != nil {
-						return nil, err
+						return nil, false, err
 					}
 				}
 			}
@@ -226,7 +224,7 @@ func readPages(ctx context.Context, r *offsetReader, col *Column, chunkMeta *par
 				ph: ph,
 			}
 		default:
-			return nil, errors.Errorf("DATA_PAGE or DATA_PAGE_V2 type supported, but was %s", ph.Type)
+			return nil, false, errors.Errorf("DATA_PAGE or DATA_PAGE_V2 type supported, but was %s", ph.Type)
 		}
 		var dictValue []interface{}
 		if dictPage != nil {
@@ -236,16 +234,16 @@ func readPages(ctx context.Context, r *offsetReader, col *Column, chunkMeta *par
 			return getValuesDecoder(typ, col.Element(), clone(dictValue))
 		}
 		if err := p.init(dDecoder, rDecoder, fn); err != nil {
-			return nil, err
+			return nil, false, err
 		}
 
 		if err := p.read(r, ph, chunkMeta.Codec); err != nil {
-			return nil, err
+			return nil, false, err
 		}
 		pages = append(pages, p)
 	}
 
-	return pages, nil
+	return pages, dictPage != nil, nil
 }
 
 func clone(in []interface{}) []interface{} {
@@ -282,9 +280,9 @@ func skipChunk(r io.Seeker, col *Column, chunk *parquet.ColumnChunk) error {
 	return err
 }
 
-func readChunk(ctx context.Context, r io.ReadSeeker, col *Column, chunk *parquet.ColumnChunk) ([]pageReader, error) {
+func readChunk(ctx context.Context, r io.ReadSeeker, col *Column, chunk *parquet.ColumnChunk) (pages []pageReader, useDict bool, err error) {
 	if chunk.FilePath != nil {
-		return nil, fmt.Errorf("nyi: data is in another file: '%s'", *chunk.FilePath)
+		return nil, false, fmt.Errorf("nyi: data is in another file: '%s'", *chunk.FilePath)
 	}
 
 	c := col.Index()
@@ -292,11 +290,11 @@ func readChunk(ctx context.Context, r io.ReadSeeker, col *Column, chunk *parquet
 	// as we cannot read it from r
 	// see https://issues.apache.org/jira/browse/PARQUET-291
 	if chunk.MetaData == nil {
-		return nil, errors.Errorf("missing meta data for Column %c", c)
+		return nil, false, errors.Errorf("missing meta data for Column %c", c)
 	}
 
 	if typ := *col.Element().Type; chunk.MetaData.Type != typ {
-		return nil, errors.Errorf("wrong type in Column chunk metadata, expected %s was %s",
+		return nil, false, errors.Errorf("wrong type in Column chunk metadata, expected %s was %s",
 			typ, chunk.MetaData.Type)
 	}
 
@@ -305,9 +303,8 @@ func readChunk(ctx context.Context, r io.ReadSeeker, col *Column, chunk *parquet
 		offset = *chunk.MetaData.DictionaryPageOffset
 	}
 	// Seek to the beginning of the first Page
-	_, err := r.Seek(offset, io.SeekStart)
-	if err != nil {
-		return nil, err
+	if _, err := r.Seek(offset, io.SeekStart); err != nil {
+		return nil, false, err
 	}
 
 	reader := &offsetReader{
@@ -348,20 +345,12 @@ func readChunk(ctx context.Context, r io.ReadSeeker, col *Column, chunk *parquet
 	return readPages(ctx, reader, col, chunk.MetaData, dDecoder, rDecoder)
 }
 
-func readPageData(col *Column, pages []pageReader) error {
+func readPageData(col *Column, pages []pageReader, useDict bool) error {
 	s := col.getColumnStore()
-	for i := range pages {
-		data, dl, rl, err := pages[i].readValues(int(pages[i].numValues()))
-		if err != nil {
-			return err
-		}
-
-		// using append to make sure we handle the multiple data page correctly
-		s.rLevels.appendArray(rl)
-		s.dLevels.appendArray(dl)
-
-		s.values.values = append(s.values.values, data...)
-		s.values.noDictMode = true
+	s.pageIdx, s.pages = 0, pages
+	s.useDict = useDict
+	if err := s.readNextPage(); err != nil {
+		return nil
 	}
 
 	return nil
@@ -384,11 +373,11 @@ func readRowGroup(ctx context.Context, r io.ReadSeeker, schema SchemaReader, row
 			c.data.skipped = true
 			continue
 		}
-		pages, err := readChunk(ctx, r, c, chunk)
+		pages, useDict, err := readChunk(ctx, r, c, chunk)
 		if err != nil {
 			return err
 		}
-		if err := readPageData(c, pages); err != nil {
+		if err := readPageData(c, pages, useDict); err != nil {
 			return err
 		}
 	}