+++ data_store.go
@@ -1,7 +1,6 @@
 package goparquet
 
 import (
-	"math"
 	"math/bits"
 
 	"github.com/fraugster/parquet-go/parquet"
@@ -17,6 +16,9 @@ type ColumnStore struct {
 
 	repTyp parquet.FieldRepetitionType
 
+	pages   []pageReader
+	pageIdx int
+
 	values *dictStore
 
 	dLevels *packedArray
@@ -25,27 +27,26 @@ type ColumnStore struct {
 	enc     parquet.Encoding
 	readPos int
 
-	allowDict bool
+	useDict bool
 
 	skipped bool
+
+	dataPages []*dataPage
+
+	maxPageSize int64
 }
 
-// useDictionary is simply a function to decide to use dictionary or not,
+type dataPage struct {
+	values     []interface{}
+	rL         *packedArray
+	dL         *packedArray
+	numValues  int64
+	nullValues int64
+}
+
+// useDictionary is simply a function to decide to use dictionary or not.
 func (cs *ColumnStore) useDictionary() bool {
-	if !cs.allowDict {
-		return false
-	}
-	if len(cs.values.data) > math.MaxInt16 {
-		return false
-	}
-
-	// There is no point for using dictionary if all values are nil
-	if len(cs.values.data) == 0 || len(cs.values.values) == 0 {
-		return false
-	}
-
-	dictLen, noDictLen := cs.values.sizes()
-	return dictLen < noDictLen
+	return cs.useDict
 }
 
 func (cs *ColumnStore) encoding() parquet.Encoding {
@@ -125,6 +126,44 @@ func (cs *ColumnStore) add(v interface{}, dL uint16, maxRL, rL uint16) error {
 	return nil
 }
 
+func (cs *ColumnStore) estimateSize() (total int64) {
+	dictSize, noDictSize := cs.values.sizes()
+	if cs.useDictionary() {
+		total += dictSize
+	} else {
+		total += noDictSize
+	}
+	total += int64(len(cs.rLevels.data) + len(cs.dLevels.data))
+	return total
+}
+
+func (cs *ColumnStore) getMaxPageSize() int64 {
+	if cs.maxPageSize == 0 {
+		return 1024 * 1024
+	}
+	return cs.maxPageSize
+}
+
+func (cs *ColumnStore) flushPage(force bool) error {
+	size := cs.estimateSize()
+
+	if !force && size < cs.getMaxPageSize() {
+		return nil
+	}
+
+	cs.dataPages = append(cs.dataPages, &dataPage{
+		values:     cs.values.getValues(),
+		rL:         cs.rLevels,
+		dL:         cs.dLevels,
+		numValues:  int64(cs.values.numValues()),
+		nullValues: int64(cs.values.nullValueCount()),
+	})
+
+	cs.resetData()
+
+	return nil
+}
+
 // getRDLevelAt return the next rLevel in the read position, if there is no value left, it returns true
 // if the position is less than zero, then it returns the current position
 // NOTE: make sure always r is before d, in any function
@@ -155,13 +194,55 @@ func (cs *ColumnStore) getNext() (v interface{}, err error) {
 	return v, nil
 }
 
+func (cs *ColumnStore) resetData() {
+	cs.readPos = 0
+	cs.values = &dictStore{}
+	cs.values.init()
+
+	rLevelBitWidth := cs.rLevels.bw
+	dLevelBitWidth := cs.dLevels.bw
+
+	cs.rLevels = &packedArray{}
+	cs.dLevels = &packedArray{}
+	cs.rLevels.reset(rLevelBitWidth)
+	cs.dLevels.reset(dLevelBitWidth)
+}
+
+func (cs *ColumnStore) readNextPage() error {
+	if cs.pageIdx >= len(cs.pages) {
+		return errors.Errorf("out of range: requested page index = %d total number of pages = %d", cs.pageIdx, len(cs.pages))
+	}
+
+	data, dl, rl, err := cs.pages[cs.pageIdx].readValues(int(cs.pages[cs.pageIdx].numValues()))
+	if err != nil {
+		return err
+	}
+
+	cs.pageIdx++
+
+	cs.resetData()
+
+	cs.values.readPos = 0
+
+	for _, v := range data {
+		cs.values.addValue(v, cs.sizeOf(v))
+	}
+
+	cs.rLevels.appendArray(rl)
+	cs.dLevels.appendArray(dl)
+
+	return nil
+}
+
 func (cs *ColumnStore) get(maxD, maxR int32) (interface{}, int32, error) {
 	if cs.skipped {
 		return nil, 0, nil
 	}
 
 	if cs.readPos >= cs.rLevels.count || cs.readPos >= cs.dLevels.count {
-		return nil, 0, errors.New("out of range")
+		if err := cs.readNextPage(); err != nil {
+			return nil, 0, err
+		}
 	}
 	_, dl, _ := cs.getRDLevelAt(cs.readPos)
 	// this is a null value, increase the read pos, for advancing the rLvl and dLvl but
@@ -202,10 +283,10 @@ func (cs *ColumnStore) get(maxD, maxR int32) (interface{}, int32, error) {
 	}
 }
 
-func newStore(typed typedColumnStore, enc parquet.Encoding, allowDict bool) *ColumnStore {
+func newStore(typed typedColumnStore, enc parquet.Encoding, useDict bool) *ColumnStore {
 	return &ColumnStore{
 		enc:              enc,
-		allowDict:        allowDict,
+		useDict:          useDict,
 		typedColumnStore: typed,
 	}
 }
@@ -264,34 +345,31 @@ func NewBooleanStore(enc parquet.Encoding, params *ColumnParameters) (*ColumnSto
 	return newStore(&booleanStore{ColumnParameters: params}, enc, false), nil
 }
 
-// NewInt32Store create a new column store to store int32 values. If allowDict is true,
-// then using a dictionary is considered by the column store depending on its heuristics.
-// If allowDict is false, a dictionary will never be used to encode the data.
-func NewInt32Store(enc parquet.Encoding, allowDict bool, params *ColumnParameters) (*ColumnStore, error) {
+// NewInt32Store create a new column store to store int32 values. If useDict is true,
+// then a dictionary is used, otherwise a dictionary will never be used to encode the data.
+func NewInt32Store(enc parquet.Encoding, useDict bool, params *ColumnParameters) (*ColumnStore, error) {
 	switch enc {
 	case parquet.Encoding_PLAIN, parquet.Encoding_DELTA_BINARY_PACKED:
 	default:
 		return nil, errors.Errorf("encoding %q is not supported on this type", enc)
 	}
-	return newStore(&int32Store{ColumnParameters: params}, enc, allowDict), nil
+	return newStore(&int32Store{ColumnParameters: params}, enc, useDict), nil
 }
 
-// NewInt64Store creates a new column store to store int64 values. If allowDict is true,
-// then using a dictionary is considered by the column store depending on its heuristics.
-// If allowDict is false, a dictionary will never be used to encode the data.
-func NewInt64Store(enc parquet.Encoding, allowDict bool, params *ColumnParameters) (*ColumnStore, error) {
+// NewInt64Store creates a new column store to store int64 values. If useDict is true,
+// then a dictionary is used, otherwise a dictionary will never be used to encode the data.
+func NewInt64Store(enc parquet.Encoding, useDict bool, params *ColumnParameters) (*ColumnStore, error) {
 	switch enc {
 	case parquet.Encoding_PLAIN, parquet.Encoding_DELTA_BINARY_PACKED:
 	default:
 		return nil, errors.Errorf("encoding %q is not supported on this type", enc)
 	}
-	return newStore(&int64Store{ColumnParameters: params}, enc, allowDict), nil
+	return newStore(&int64Store{ColumnParameters: params}, enc, useDict), nil
 }
 
-// NewInt96Store creates a new column store to store int96 values. If allowDict is true,
-// then using a dictionary is considered by the column store depending on its heuristics.
-// If allowDict is false, a dictionary will never be used to encode the data.
-func NewInt96Store(enc parquet.Encoding, allowDict bool, params *ColumnParameters) (*ColumnStore, error) {
+// NewInt96Store creates a new column store to store int96 values. If useDict is true,
+// then a dictionary is used, otherwise a dictionary will never be used to encode the data.
+func NewInt96Store(enc parquet.Encoding, useDict bool, params *ColumnParameters) (*ColumnStore, error) {
 	switch enc {
 	case parquet.Encoding_PLAIN:
 	default:
@@ -299,49 +377,45 @@ func NewInt96Store(enc parquet.Encoding, allowDict bool, params *ColumnParameter
 	}
 	store := &int96Store{}
 	store.ColumnParameters = params
-	return newStore(store, enc, allowDict), nil
+	return newStore(store, enc, useDict), nil
 }
 
-// NewFloatStore creates a new column store to store float (float32) values. If allowDict is true,
-// then using a dictionary is considered by the column store depending on its heuristics.
-// If allowDict is false, a dictionary will never be used to encode the data.
-func NewFloatStore(enc parquet.Encoding, allowDict bool, params *ColumnParameters) (*ColumnStore, error) {
+// NewFloatStore creates a new column store to store float (float32) values. If useDict is true,
+// then a dictionary is used, otherwise a dictionary will never be used to encode the data.
+func NewFloatStore(enc parquet.Encoding, useDict bool, params *ColumnParameters) (*ColumnStore, error) {
 	switch enc {
 	case parquet.Encoding_PLAIN:
 	default:
 		return nil, errors.Errorf("encoding %q is not supported on this type", enc)
 	}
-	return newStore(&floatStore{ColumnParameters: params}, enc, allowDict), nil
+	return newStore(&floatStore{ColumnParameters: params}, enc, useDict), nil
 }
 
-// NewDoubleStore creates a new column store to store double (float64) values. If allowDict is true,
-// then using a dictionary is considered by the column store depending on its heuristics.
-// If allowDict is false, a dictionary will never be used to encode the data.
-func NewDoubleStore(enc parquet.Encoding, allowDict bool, params *ColumnParameters) (*ColumnStore, error) {
+// NewDoubleStore creates a new column store to store double (float64) values. If useDict is true,
+// then a dictionary is used, otherwise a dictionary will never be used to encode the data.
+func NewDoubleStore(enc parquet.Encoding, useDict bool, params *ColumnParameters) (*ColumnStore, error) {
 	switch enc {
 	case parquet.Encoding_PLAIN:
 	default:
 		return nil, errors.Errorf("encoding %q is not supported on this type", enc)
 	}
-	return newStore(&doubleStore{ColumnParameters: params}, enc, allowDict), nil
+	return newStore(&doubleStore{ColumnParameters: params}, enc, useDict), nil
 }
 
-// NewByteArrayStore creates a new column store to store byte arrays. If allowDict is true,
-// then using a dictionary is considered by the column store depending on its heuristics.
-// If allowDict is false, a dictionary will never be used to encode the data.
-func NewByteArrayStore(enc parquet.Encoding, allowDict bool, params *ColumnParameters) (*ColumnStore, error) {
+// NewByteArrayStore creates a new column store to store byte arrays. If useDict is true,
+// then a dictionary is used, otherwise a dictionary will never be used to encode the data.
+func NewByteArrayStore(enc parquet.Encoding, useDict bool, params *ColumnParameters) (*ColumnStore, error) {
 	switch enc {
 	case parquet.Encoding_PLAIN, parquet.Encoding_DELTA_LENGTH_BYTE_ARRAY, parquet.Encoding_DELTA_BYTE_ARRAY:
 	default:
 		return nil, errors.Errorf("encoding %q is not supported on this type", enc)
 	}
-	return newStore(&byteArrayStore{ColumnParameters: params}, enc, allowDict), nil
+	return newStore(&byteArrayStore{ColumnParameters: params}, enc, useDict), nil
 }
 
-// NewFixedByteArrayStore creates a new column store to store fixed size byte arrays. If allowDict is true,
-// then using a dictionary is considered by the column store depending on its heuristics.
-// If allowDict is false, a dictionary will never be used to encode the data.
-func NewFixedByteArrayStore(enc parquet.Encoding, allowDict bool, params *ColumnParameters) (*ColumnStore, error) {
+// NewFixedByteArrayStore creates a new column store to store fixed size byte arrays. If useDict is true,
+// then a dictionary is used, otherwise a dictionary will never be used to encode the data.
+func NewFixedByteArrayStore(enc parquet.Encoding, useDict bool, params *ColumnParameters) (*ColumnStore, error) {
 	switch enc {
 	case parquet.Encoding_PLAIN, parquet.Encoding_DELTA_LENGTH_BYTE_ARRAY, parquet.Encoding_DELTA_BYTE_ARRAY:
 	default:
@@ -357,5 +431,5 @@ func NewFixedByteArrayStore(enc parquet.Encoding, allowDict bool, params *Column
 
 	return newStore(&byteArrayStore{
 		ColumnParameters: params,
-	}, enc, allowDict), nil
+	}, enc, useDict), nil
 }