+++ data_store_test.go
@@ -32,7 +32,7 @@ func TestOneColumn(t *testing.T) {
 	require.NoError(t, err)
 	assert.Equal(t, uint16(0), d.MaxDefinitionLevel())
 	assert.Equal(t, uint16(0), d.MaxRepetitionLevel())
-	assert.Equal(t, []interface{}{int32(10), int32(20)}, d.data.values.assemble())
+	assert.Equal(t, []interface{}{int32(10), int32(20)}, d.data.values.getValues())
 	assert.Equal(t, []int32{0, 0}, d.data.dLevels.toArray())
 	assert.Equal(t, []int32{0, 0}, d.data.rLevels.toArray())
 
@@ -62,7 +62,7 @@ func TestOneColumnOptional(t *testing.T) {
 	require.NoError(t, err)
 	assert.Equal(t, uint16(1), d.MaxDefinitionLevel())
 	assert.Equal(t, uint16(0), d.MaxRepetitionLevel())
-	assert.Equal(t, []interface{}{int32(10)}, d.data.values.assemble())
+	assert.Equal(t, []interface{}{int32(10)}, d.data.values.getValues())
 	assert.Equal(t, []int32{1, 0}, d.data.dLevels.toArray())
 	assert.Equal(t, []int32{0, 0}, d.data.rLevels.toArray())
 
@@ -90,7 +90,7 @@ func TestOneColumnRepeated(t *testing.T) {
 	require.NoError(t, err)
 	assert.Equal(t, uint16(1), d.MaxDefinitionLevel())
 	assert.Equal(t, uint16(1), d.MaxRepetitionLevel())
-	assert.Equal(t, []interface{}{int32(10), int32(20)}, d.data.values.assemble())
+	assert.Equal(t, []interface{}{int32(10), int32(20)}, d.data.values.getValues())
 	assert.Equal(t, []int32{1, 1, 0}, d.data.dLevels.toArray())
 	assert.Equal(t, []int32{0, 1, 0}, d.data.rLevels.toArray())
 
@@ -149,7 +149,7 @@ func TestComplexPart1(t *testing.T) {
 	require.NoError(t, err)
 	assert.Equal(t, uint16(2), d.MaxDefinitionLevel())
 	assert.Equal(t, uint16(2), d.MaxRepetitionLevel())
-	assert.Equal(t, []interface{}{int32(1), int32(2), int32(3)}, d.data.values.assemble())
+	assert.Equal(t, []interface{}{int32(1), int32(2), int32(3)}, d.data.values.getValues())
 	assert.Equal(t, []int32{2, 2, 1, 2}, d.data.dLevels.toArray())
 	assert.Equal(t, []int32{0, 2, 1, 1}, d.data.rLevels.toArray())
 
@@ -157,7 +157,7 @@ func TestComplexPart1(t *testing.T) {
 	require.NoError(t, err)
 	assert.Equal(t, uint16(3), d.MaxDefinitionLevel())
 	assert.Equal(t, uint16(2), d.MaxRepetitionLevel())
-	assert.Equal(t, []interface{}{int32(100), int32(101)}, d.data.values.assemble())
+	assert.Equal(t, []interface{}{int32(100), int32(101)}, d.data.values.getValues())
 	assert.Equal(t, []int32{3, 2, 1, 3}, d.data.dLevels.toArray())
 	assert.Equal(t, []int32{0, 2, 1, 1}, d.data.rLevels.toArray())
 
@@ -165,7 +165,7 @@ func TestComplexPart1(t *testing.T) {
 	require.NoError(t, err)
 	assert.Equal(t, uint16(2), d.MaxDefinitionLevel())
 	assert.Equal(t, uint16(1), d.MaxRepetitionLevel())
-	assert.Equal(t, []interface{}{int32(10), int32(11)}, d.data.values.assemble())
+	assert.Equal(t, []interface{}{int32(10), int32(11)}, d.data.values.getValues())
 	assert.Equal(t, []int32{2, 2, 1}, d.data.dLevels.toArray())
 	assert.Equal(t, []int32{0, 1, 1}, d.data.rLevels.toArray())
 
@@ -205,7 +205,7 @@ func TestComplexPart2(t *testing.T) {
 	require.NoError(t, err)
 	assert.Equal(t, uint16(2), d.MaxDefinitionLevel())
 	assert.Equal(t, uint16(1), d.MaxRepetitionLevel())
-	assert.Equal(t, []interface{}{int32(20), int32(40), int32(60), int32(80)}, d.data.values.assemble())
+	assert.Equal(t, []interface{}{int32(20), int32(40), int32(60), int32(80)}, d.data.values.getValues())
 	assert.Equal(t, []int32{2, 2, 2, 2}, d.data.dLevels.toArray())
 	assert.Equal(t, []int32{0, 1, 1, 0}, d.data.rLevels.toArray())
 
@@ -213,7 +213,7 @@ func TestComplexPart2(t *testing.T) {
 	require.NoError(t, err)
 	assert.Equal(t, uint16(2), d.MaxDefinitionLevel())
 	assert.Equal(t, uint16(1), d.MaxRepetitionLevel())
-	assert.Equal(t, []interface{}{int32(10), int32(30)}, d.data.values.assemble())
+	assert.Equal(t, []interface{}{int32(10), int32(30)}, d.data.values.getValues())
 	assert.Equal(t, []int32{1, 2, 2}, d.data.dLevels.toArray())
 	assert.Equal(t, []int32{0, 0, 1}, d.data.rLevels.toArray())
 
@@ -292,7 +292,7 @@ func TestComplex(t *testing.T) {
 	require.NoError(t, err)
 	assert.Equal(t, uint16(0), d.MaxDefinitionLevel())
 	assert.Equal(t, uint16(0), d.MaxRepetitionLevel())
-	assert.Equal(t, []interface{}{int32(10), int32(20)}, d.data.values.assemble())
+	assert.Equal(t, []interface{}{int32(10), int32(20)}, d.data.values.getValues())
 	assert.Equal(t, []int32{0, 0}, d.data.dLevels.toArray())
 	assert.Equal(t, []int32{0, 0}, d.data.rLevels.toArray())
 
@@ -300,7 +300,7 @@ func TestComplex(t *testing.T) {
 	require.NoError(t, err)
 	assert.Equal(t, uint16(2), d.MaxDefinitionLevel())
 	assert.Equal(t, uint16(1), d.MaxRepetitionLevel())
-	assert.Equal(t, []interface{}{int32(10), int32(11), int32(12)}, d.data.values.assemble())
+	assert.Equal(t, []interface{}{int32(10), int32(11), int32(12)}, d.data.values.getValues())
 	assert.Equal(t, []int32{2, 2, 1, 2}, d.data.dLevels.toArray())
 	assert.Equal(t, []int32{0, 1, 1, 0}, d.data.rLevels.toArray())
 
@@ -308,7 +308,7 @@ func TestComplex(t *testing.T) {
 	require.NoError(t, err)
 	assert.Equal(t, uint16(2), d.MaxDefinitionLevel())
 	assert.Equal(t, uint16(1), d.MaxRepetitionLevel())
-	assert.Equal(t, []interface{}{int32(20), int32(40), int32(60), int32(80)}, d.data.values.assemble())
+	assert.Equal(t, []interface{}{int32(20), int32(40), int32(60), int32(80)}, d.data.values.getValues())
 	assert.Equal(t, []int32{2, 2, 2, 2}, d.data.dLevels.toArray())
 	assert.Equal(t, []int32{0, 1, 1, 0}, d.data.rLevels.toArray())
 
@@ -316,7 +316,7 @@ func TestComplex(t *testing.T) {
 	require.NoError(t, err)
 	assert.Equal(t, uint16(2), d.MaxDefinitionLevel())
 	assert.Equal(t, uint16(1), d.MaxRepetitionLevel())
-	assert.Equal(t, []interface{}{int32(10), int32(30)}, d.data.values.assemble())
+	assert.Equal(t, []interface{}{int32(10), int32(30)}, d.data.values.getValues())
 	assert.Equal(t, []int32{1, 2, 2}, d.data.dLevels.toArray())
 	assert.Equal(t, []int32{0, 0, 1}, d.data.rLevels.toArray())
 
@@ -324,7 +324,7 @@ func TestComplex(t *testing.T) {
 	require.NoError(t, err)
 	assert.Equal(t, uint16(3), d.MaxDefinitionLevel())
 	assert.Equal(t, uint16(2), d.MaxRepetitionLevel())
-	assert.Equal(t, []interface{}{int32(100), int32(101)}, d.data.values.assemble())
+	assert.Equal(t, []interface{}{int32(100), int32(101)}, d.data.values.getValues())
 	assert.Equal(t, []int32{3, 2, 1, 3, 1}, d.data.dLevels.toArray())
 	assert.Equal(t, []int32{0, 2, 1, 1, 0}, d.data.rLevels.toArray())
 
@@ -332,7 +332,7 @@ func TestComplex(t *testing.T) {
 	require.NoError(t, err)
 	assert.Equal(t, uint16(2), d.MaxDefinitionLevel())
 	assert.Equal(t, uint16(2), d.MaxRepetitionLevel())
-	assert.Equal(t, []interface{}{int32(1), int32(2), int32(3)}, d.data.values.assemble())
+	assert.Equal(t, []interface{}{int32(1), int32(2), int32(3)}, d.data.values.getValues())
 	assert.Equal(t, []int32{2, 2, 1, 2, 1}, d.data.dLevels.toArray())
 	assert.Equal(t, []int32{0, 2, 1, 1, 0}, d.data.rLevels.toArray())
 
@@ -375,7 +375,7 @@ func TestTwitterBlog(t *testing.T) {
 	for i := 1; i < 11; i++ {
 		expected = append(expected, int32(i))
 	}
-	assert.Equal(t, expected, d.data.values.assemble())
+	assert.Equal(t, expected, d.data.values.getValues())
 	assert.Equal(t, uint16(2), d.MaxDefinitionLevel())
 	assert.Equal(t, uint16(2), d.MaxRepetitionLevel())
 	assert.Equal(t, []int32{0, 2, 2, 1, 2, 2, 2, 0, 1, 2}, d.data.rLevels.toArray())
@@ -412,7 +412,7 @@ func TestEmptyParent(t *testing.T) {
 	col, err := row.findDataColumn("baz.list.element")
 	require.NoError(t, err)
 
-	assert.Equal(t, []interface{}{}, col.data.values.assemble())
+	assert.Equal(t, []interface{}(nil), col.data.values.getValues())
 
 	assert.Equal(t, uint16(2), col.MaxDefinitionLevel())
 	assert.Equal(t, uint16(1), col.MaxRepetitionLevel())
@@ -465,7 +465,7 @@ func TestZeroRL(t *testing.T) {
 	d, err := row.findDataColumn("baz.list.element.quux")
 	require.NoError(t, err)
 	var expected = []interface{}{int32(23), int32(42)}
-	assert.Equal(t, expected, d.data.values.assemble())
+	assert.Equal(t, expected, d.data.values.getValues())
 	assert.Equal(t, uint16(1), d.MaxDefinitionLevel())
 	assert.Equal(t, uint16(1), d.MaxRepetitionLevel())
 	assert.Equal(t, []int32{0, 1}, d.data.rLevels.toArray())
@@ -485,7 +485,7 @@ func TestZeroRL(t *testing.T) {
 
 	d, err = row.findDataColumn("baz.list.element.quux")
 	require.NoError(t, err)
-	assert.Equal(t, expected, d.data.values.assemble())
+	assert.Equal(t, expected, d.data.values.getValues())
 	assert.Equal(t, uint16(2), d.MaxDefinitionLevel())
 	assert.Equal(t, uint16(1), d.MaxRepetitionLevel())
 	assert.Equal(t, []int32{0, 1}, d.data.rLevels.toArray())