+++ readwrite_test.go
@@ -5,6 +5,7 @@ import (
 	"context"
 	"fmt"
 	"io"
+	"math/rand"
 	"os"
 	"testing"
 	"time"
@@ -19,7 +20,7 @@ import (
 func TestWriteThenReadFile(t *testing.T) {
 	ctx := context.Background()
 
-	testFunc := func(opts ...FileWriterOption) {
+	testFunc := func(t *testing.T, opts ...FileWriterOption) {
 		_ = os.Mkdir("files", 0755)
 
 		wf, err := os.OpenFile("files/test1.parquet", os.O_CREATE|os.O_TRUNC|os.O_WRONLY, 0644)
@@ -61,7 +62,7 @@ func TestWriteThenReadFile(t *testing.T) {
 		require.NoError(t, err, "creating file reader failed")
 
 		cols := r.Columns()
-		require.Len(t, cols, 2, fmt.Sprintf("expected 2 columns, got %d instead", len(cols)))
+		require.Len(t, cols, 2, "got %d column", len(cols))
 		require.Equal(t, "foo", cols[0].Name())
 		require.Equal(t, "foo", cols[0].FlatName())
 		require.Equal(t, "bar", cols[1].Name())
@@ -77,8 +78,12 @@ func TestWriteThenReadFile(t *testing.T) {
 		}
 	}
 
-	testFunc(WithCompressionCodec(parquet.CompressionCodec_SNAPPY), WithCreator("parquet-go-unittest"))
-	testFunc(WithCompressionCodec(parquet.CompressionCodec_SNAPPY), WithCreator("parquet-go-unittest"), WithDataPageV2())
+	t.Run("datapagev1", func(t *testing.T) {
+		testFunc(t, WithCompressionCodec(parquet.CompressionCodec_SNAPPY), WithCreator("parquet-go-unittest"))
+	})
+	t.Run("datapagev2", func(t *testing.T) {
+		testFunc(t, WithCompressionCodec(parquet.CompressionCodec_SNAPPY), WithCreator("parquet-go-unittest"), WithDataPageV2())
+	})
 }
 
 func TestWriteThenReadFileRepeated(t *testing.T) {
@@ -1072,4 +1077,44 @@ func TestReadWriteDeltaBinaryPackedInt64(t *testing.T) {
 	require.True(t, errors.Is(err, io.EOF))
 
 	t.Logf("Finished")
+}
+
+func TestWriteThenReadMultiplePages(t *testing.T) {
+	const mySchema = `message msg {
+		required binary ts_str (STRING);
+	}`
+
+	sd, err := parquetschema.ParseSchemaDefinition(mySchema)
+	require.NoError(t, err)
+
+	f := new(bytes.Buffer)
+
+	fw := NewFileWriter(f, WithSchemaDefinition(sd), WithCompressionCodec(parquet.CompressionCodec_SNAPPY))
+	defer fw.Close()
+
+	const numRows = 100000
+
+	records := []map[string]interface{}{}
+
+	for i := 0; i < numRows; i++ {
+		tsStr := time.Now().Add(time.Duration(1+rand.Int63n(300)) * time.Second).Format(time.RFC3339)
+		rec := map[string]interface{}{"ts_str": []byte(tsStr)}
+		records = append(records, rec)
+		require.NoError(t, fw.AddData(rec))
+	}
+
+	require.NoError(t, fw.Close())
+
+	r, err := NewFileReader(bytes.NewReader(f.Bytes()))
+	require.NoError(t, err)
+
+	rowCount := r.NumRows()
+	require.Equal(t, int64(numRows), rowCount)
+
+	for i := int64(0); i < rowCount; i++ {
+		data, err := r.NextRow()
+		require.NoError(t, err)
+		require.Equal(t, records[i], data, "%d. records don't match", i)
+		//fmt.Printf("in %d. %s\n", i, string(data["ts_str"].([]byte)))
+	}
 }